+++ index.js
@@ -1,5 +1,5 @@
 var next = require('./lib/index.js');
 
-next.version = '1.11.3';
+next.version = '1.11.5';
 
 module.exports = next;